pyproject.toml
[project]
name = "ACBFEditor"
#dynamic = ["version"]
version = "3.0.0a"
authors = [
  {name = "Robert Kubik"},
  {name = "mizaki"},
]
description = "Editor for comic book files in ACBF format"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["comics", "comic", "metadata", "tagging", "tagger"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Other/Nonlisted Topic",
    "Topic :: Utilities",
    "License :: OSI Approved :: GPLv3 License",
]
requires-python = ">= 3.9"
dependencies = [
    "PyGObject",
    "lxml",
    "pillow>10",
    "opencv-python",
    "patool",
    "pycountry",
]
[project.optional-dependencies]
dev = [
  "pre-commit",
  "ruff",
  "mypy",
]

[project.urls]
Homepage = "https://github.com/GeoRW/ACBF-Editor"
Documentation = "https://github.com/GeoRW/ACBF-Editor/wiki"
Repository = "https://github.com/GeoRW/ACBF-Editor.git"
Issues = "https://github.com/GeoRW/ACBF-Editor/issues"

[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=3.4"]
build-backend = "setuptools.build_meta"

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
warn_return_any = false
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
exclude = [
    "kumiko",
]

[[tool.mypy.overrides]]
module = ["kumiko.*"]
follow_imports = "skip"

[tool.ruff]
line-length = 120
lint.extend-safe-fixes = ["TCH"]
lint.extend-select = ["TCH"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "kumiko",
]

[tool.ruff.format]
